+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "lint": "standard | snazzy",
-    "unit": "tap test.js",
+    "unit": "tap test/*test.js",
     "test": "npm run lint && npm run unit && npm run typescript",
     "typescript": "tsd"
   },
