+++ test/test.js
@@ -5,7 +5,7 @@ const { promisify } = require('util')
 const sget = require('simple-get').concat
 const Fastify = require('fastify')
 const { monitorEventLoopDelay } = require('perf_hooks')
-const underPressure = require('./index')
+const underPressure = require('../index')
 const { valid, satisfies, coerce } = require('semver')
 
 const wait = promisify(setTimeout)
@@ -333,185 +333,6 @@ test('memoryUsage name space (without check)', t => {
   })
 })
 
-test('Expose status route', t => {
-  t.plan(4)
-
-  const fastify = Fastify()
-  fastify.register(underPressure, {
-    exposeStatusRoute: true
-  })
-
-  fastify.listen(0, (err, address) => {
-    t.error(err)
-    fastify.server.unref()
-
-    process.nextTick(() => block(monitorEventLoopDelay ? 1500 : 500))
-
-    sget({
-      method: 'GET',
-      url: `${address}/status`
-    }, (err, response, body) => {
-      t.error(err)
-      t.strictEqual(response.statusCode, 200)
-      t.deepEqual(JSON.parse(body), { status: 'ok' })
-      fastify.close()
-    })
-  })
-})
-
-test('Expose custom status route', t => {
-  t.plan(5)
-
-  const fastify = Fastify()
-  t.tearDown(() => fastify.close())
-
-  fastify.register(underPressure, {
-    exposeStatusRoute: '/alive'
-  })
-
-  fastify.inject({
-    url: '/status'
-  }, (err, response) => {
-    t.error(err)
-    t.strictEqual(response.statusCode, 404)
-  })
-
-  fastify.inject({
-    url: '/alive'
-  }, (err, response) => {
-    t.error(err)
-    t.strictEqual(response.statusCode, 200)
-    t.deepEqual(JSON.parse(response.payload), { status: 'ok' })
-  })
-})
-
-test('Expose status route with additional route options', t => {
-  t.plan(3)
-
-  const customConfig = {
-    customVal: 'someVal'
-  }
-  const fastify = Fastify()
-
-  fastify.addHook('onRoute', (routeOptions) => {
-    fastify.server.unref()
-    process.nextTick(() => block(500))
-    t.strictEqual(routeOptions.url, '/alive')
-    t.strictEqual(routeOptions.logLevel, 'silent', 'log level not set')
-    t.deepEqual(routeOptions.config, customConfig, 'config not set')
-    fastify.close()
-  })
-
-  fastify.register(underPressure, {
-    exposeStatusRoute: {
-      routeOpts: {
-        logLevel: 'silent',
-        config: customConfig
-      },
-      url: '/alive'
-    }
-  })
-
-  fastify.ready()
-})
-
-test('Expose status route with additional route options and default url', t => {
-  t.plan(2)
-
-  const fastify = Fastify()
-
-  fastify.addHook('onRoute', (routeOptions) => {
-    fastify.server.unref()
-    process.nextTick(() => block(500))
-    t.strictEqual(routeOptions.url, '/status')
-    t.strictEqual(routeOptions.logLevel, 'silent', 'log level not set')
-    fastify.close()
-  })
-
-  fastify.register(underPressure, {
-    exposeStatusRoute: {
-      routeOpts: {
-        logLevel: 'silent'
-      }
-    }
-  })
-
-  fastify.ready()
-})
-
-test('Expose status route with additional route options, route schema options', t => {
-  const routeSchemaOpts = { hide: true }
-
-  const fastify = Fastify()
-
-  fastify.addHook('onRoute', (routeOptions) => {
-    fastify.server.unref()
-    process.nextTick(() => block(500))
-    t.strictEqual(routeOptions.url, '/alive')
-    t.strictEqual(routeOptions.logLevel, 'silent', 'log level not set')
-    t.deepEqual(routeOptions.schema, Object.assign({}, routeSchemaOpts, {
-      response: {
-        200: {
-          type: 'object',
-          properties: {
-            status: { type: 'string' }
-          }
-        }
-      }
-    }), 'config not set')
-    fastify.close()
-    t.end()
-  })
-
-  fastify.register(underPressure, {
-    exposeStatusRoute: {
-      routeOpts: {
-        logLevel: 'silent'
-      },
-      routeSchemaOpts,
-      url: '/alive'
-    }
-  })
-
-  fastify.ready()
-})
-
-test('Expose status route with additional route options, route schema options and default url', t => {
-  const routeSchemaOpts = { hide: true }
-
-  const fastify = Fastify()
-
-  fastify.addHook('onRoute', (routeOptions) => {
-    fastify.server.unref()
-    process.nextTick(() => block(500))
-    t.strictEqual(routeOptions.url, '/status')
-    t.strictEqual(routeOptions.logLevel, 'silent', 'log level not set')
-    t.deepEqual(routeOptions.schema, Object.assign({}, routeSchemaOpts, {
-      response: {
-        200: {
-          type: 'object',
-          properties: {
-            status: { type: 'string' }
-          }
-        }
-      }
-    }), 'config not set')
-    fastify.close()
-    t.end()
-  })
-
-  fastify.register(underPressure, {
-    exposeStatusRoute: {
-      routeOpts: {
-        logLevel: 'silent'
-      },
-      routeSchemaOpts
-    }
-  })
-
-  fastify.ready()
-})
-
 test('Custom health check', t => {
   t.plan(7)
 
@@ -765,228 +586,6 @@ test('Custom health check', t => {
           anotherValue: 'another',
           status: 'overrride status'
         })
-        fastify.close()
-      })
-    })
-  })
-})
-
-test('Pressure handler', t => {
-  t.plan(9)
-
-  t.test('health check', async t => {
-    t.plan(3)
-    const fastify = Fastify()
-
-    fastify.register(underPressure, {
-      healthCheck: async () => false,
-      healthCheckInterval: 1,
-      pressureHandler: (req, rep, type, value) => {
-        t.equal(type, underPressure.TYPE_HEALTH_CHECK)
-        t.equal(value, undefined)
-        rep.send('B')
-      }
-    })
-
-    fastify.get('/', (req, rep) => rep.send('A'))
-
-    t.equal((await fastify.inject().get('/').end()).body, 'B')
-  })
-
-  t.test('health check - delayed handling with promise success', async t => {
-    t.plan(1)
-    const fastify = Fastify()
-
-    fastify.register(underPressure, {
-      healthCheck: async () => false,
-      healthCheckInterval: 1,
-      pressureHandler: async (req, rep, type, value) => {
-        await wait(250)
-        rep.send('B')
-      }
-    })
-
-    fastify.get('/', (req, rep) => rep.send('A'))
-
-    t.equal((await fastify.inject().get('/').end()).body, 'B')
-  })
-
-  t.test('health check - delayed handling with promise error', async t => {
-    t.plan(2)
-    const fastify = Fastify()
-
-    const errorMessage = 'promiseError'
-
-    fastify.register(underPressure, {
-      healthCheck: async () => false,
-      healthCheckInterval: 1,
-      pressureHandler: async (req, rep, type, value) => {
-        await wait(250)
-        throw new Error(errorMessage)
-      }
-    })
-
-    fastify.get('/', (req, rep) => rep.send('A'))
-
-    const response = await fastify.inject().get('/').end()
-    t.equal(response.statusCode, 500)
-    t.equal(JSON.parse(response.body).message, errorMessage)
-  })
-
-  t.test('health check - no handling', async t => {
-    t.plan(1)
-    const fastify = Fastify()
-
-    fastify.register(underPressure, {
-      healthCheck: async () => false,
-      healthCheckInterval: 1,
-      pressureHandler: (req, rep, type, value) => {}
-    })
-
-    fastify.get('/', (req, rep) => rep.send('A'))
-
-    t.equal((await fastify.inject().get('/').end()).body, 'A')
-  })
-
-  t.test('health check - return response', async t => {
-    t.plan(1)
-    const fastify = Fastify()
-
-    fastify.register(underPressure, {
-      healthCheck: async () => false,
-      healthCheckInterval: 1,
-      pressureHandler: (req, rep, type, value) => 'B'
-    })
-
-    fastify.get('/', (req, rep) => rep.send('A'))
-
-    t.equal((await fastify.inject().get('/').end()).body, 'B')
-  })
-
-  t.test('event loop delay', { skip: !monitorEventLoopDelay }, t => {
-    t.plan(5)
-    const fastify = Fastify()
-
-    fastify.register(underPressure, {
-      maxEventLoopDelay: 1,
-      pressureHandler: (req, rep, type, value) => {
-        t.equal(type, underPressure.TYPE_EVENT_LOOP_DELAY)
-        t.ok(value > 1)
-        rep.send('B')
-      }
-    })
-
-    fastify.get('/', (req, rep) => rep.send('A'))
-
-    fastify.listen(0, async (err, address) => {
-      t.error(err)
-      fastify.server.unref()
-
-      await wait(500)
-      process.nextTick(() => block(1500))
-
-      sget({
-        method: 'GET',
-        url: address + '/'
-      }, (err, response, body) => {
-        t.error(err)
-        t.equal(body.toString(), 'B')
-        fastify.close()
-      })
-    })
-  })
-
-  t.test('heap bytes', t => {
-    t.plan(5)
-
-    const fastify = Fastify()
-    fastify.register(underPressure, {
-      maxHeapUsedBytes: 1,
-      pressureHandler: (req, rep, type, value) => {
-        t.equal(type, underPressure.TYPE_HEAP_USED_BYTES)
-        t.ok(value > 1)
-        rep.send('B')
-      }
-    })
-
-    fastify.get('/', (req, rep) => rep.send('A'))
-
-    fastify.listen(0, (err, address) => {
-      t.error(err)
-      fastify.server.unref()
-
-      process.nextTick(() => block(monitorEventLoopDelay ? 1500 : 500))
-
-      sget({
-        method: 'GET',
-        url: address
-      }, (err, response, body) => {
-        t.error(err)
-        t.equal(body.toString(), 'B')
-        fastify.close()
-      })
-    })
-  })
-
-  t.test('rss bytes', t => {
-    t.plan(5)
-
-    const fastify = Fastify()
-    fastify.register(underPressure, {
-      maxRssBytes: 1,
-      pressureHandler: (req, rep, type, value) => {
-        t.equal(type, underPressure.TYPE_RSS_BYTES)
-        t.ok(value > 1)
-        rep.send('B')
-      }
-    })
-
-    fastify.get('/', (req, rep) => rep.send('A'))
-
-    fastify.listen(0, (err, address) => {
-      t.error(err)
-      fastify.server.unref()
-
-      process.nextTick(() => block(monitorEventLoopDelay ? 1500 : 500))
-
-      sget({
-        method: 'GET',
-        url: address
-      }, (err, response, body) => {
-        t.error(err)
-        t.equal(body.toString(), 'B')
-        fastify.close()
-      })
-    })
-  })
-
-  t.test('event loop utilization', { skip: !isSupportedVersion }, t => {
-    t.plan(5)
-
-    const fastify = Fastify()
-    fastify.register(underPressure, {
-      maxEventLoopUtilization: 0.01,
-      pressureHandler: (req, rep, type, value) => {
-        t.equal(type, underPressure.TYPE_EVENT_LOOP_UTILIZATION)
-        t.ok(value > 0.01 && value <= 1)
-        rep.send('B')
-      }
-    })
-
-    fastify.get('/', async (req, rep) => rep.send('A'))
-
-    fastify.listen(0, (err, address) => {
-      t.error(err)
-      fastify.server.unref()
-
-      process.nextTick(() => block(1000))
-
-      sget({
-        method: 'GET',
-        url: address
-      }, (err, response, body) => {
-        t.error(err)
-        t.equal(body.toString(), 'B')
         fastify.close()
       })
     })
